+++ src/main/java/com/exadel/project/HealthCheckController.java
@@ -6,12 +6,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 
-
 @RestController
 @RequestMapping("/api/example")
 public class HealthCheckController {
-
-
 
     @GetMapping("/check")
     public ResponseEntity<String> get() {
