+++ src/main/java/com/exadel/project/administrator/entity/RoleAdmin.java
@@ -1,4 +1,4 @@
-package com.exadel.project.entity;
+package com.exadel.project.administrator.entity;
 
 public enum RoleAdmin {
     ADMIN, SUPERADMIN
