--- src/main/java/com/exadel/project/entity/TypeInterviewer.java
@@ -1,5 +1,0 @@
-package com.exadel.project.entity;
-
-public enum TypeInterviewer {
-    HR, TECH
-}
