+++ src/main/java/com/exadel/project/generic_service/GenericRepositoryServiceImpl.java
@@ -1,4 +1,4 @@
-package com.exadel.project.service;
+package com.exadel.project.generic_service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
