+++ src/main/java/com/exadel/project/interviewer/entity/TypeInterviewer.java
@@ -1,0 +1,5 @@
+package com.exadel.project.interviewer.entity;
+
+public enum TypeInterviewer {
+    HR, TECH
+}
