--- src/main/java/com/exadel/project/service/TraineeRepository.java
@@ -1,8 +1,0 @@
-package com.exadel.project.service;
-
-import com.exadel.project.entity.Trainee;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface TraineeRepository extends GenericRepository<Trainee,Long> {
-}
