--- src/main/java/com/exadel/project/service/TraineeRepositoryService.java
@@ -1,9 +1,0 @@
-package com.exadel.project.service;
-
-import com.exadel.project.entity.Trainee;
-
-public interface TraineeRepositoryService extends GenericRepositoryService<Trainee, Long>  {
-
-
-
-}
