--- src/main/java/com/exadel/project/service/TraineeRepositoryServiceImpl.java
@@ -1,7 +1,0 @@
-package com.exadel.project.service;
-
-import com.exadel.project.entity.Trainee;
-
-public class TraineeRepositoryServiceImpl extends GenericRepositoryServiceImpl<Trainee, Long>
-        implements TraineeRepositoryService {
-}
