+++ src/main/java/com/exadel/project/trainee/service/TraineeRepository.java
@@ -1,0 +1,9 @@
+package com.exadel.project.trainee.service;
+
+import com.exadel.project.generic_service.GenericRepository;
+import com.exadel.project.trainee.entity.Trainee;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface TraineeRepository extends GenericRepository<Trainee,Long> {
+}
