+++ src/main/java/com/exadel/project/trainee/service/TraineeRepositoryService.java
@@ -1,0 +1,10 @@
+package com.exadel.project.trainee.service;
+
+import com.exadel.project.generic_service.GenericRepositoryService;
+import com.exadel.project.trainee.entity.Trainee;
+
+public interface TraineeRepositoryService extends GenericRepositoryService<Trainee, Long> {
+
+
+
+}
