+++ src/main/java/com/exadel/project/trainee/service/TraineeRepositoryServiceImpl.java
@@ -1,0 +1,8 @@
+package com.exadel.project.trainee.service;
+
+import com.exadel.project.generic_service.GenericRepositoryServiceImpl;
+import com.exadel.project.trainee.entity.Trainee;
+
+public class TraineeRepositoryServiceImpl extends GenericRepositoryServiceImpl<Trainee, Long>
+        implements TraineeRepositoryService {
+}
